requirements.txt
python > 3.5
pandas
xlrd
msoffcrypto-tool
